Revamped the build build process
This PR moves all files from the `src` folder to the root so it matches what is done with other `core` modules and makes it easier for contributors to find their way around this procject.

Cake was having issues myteriously and instead of going down that rabbit hole, I migrated the build process to Nuke which has way less breaking changes and is way easier to understand.